pyproject.toml
[tool.poetry]
name = "umgpt-add-on"
version = "0.1.0"
description = ""
authors = ["astral <astral@astral.cafe>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
openai = "^1.47.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.30.6"
google-auth = "^2.35.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
